requirements.txt
# ==================== CORE FRAMEWORK ====================
fastapi>=0.104.0,<1.0.0          # Web framework (async)
uvicorn[standard]>=0.24.0        # ASGI server
jinja2>=3.1.0                    # HTML template engine
python-multipart>=0.0.6          # Required for UploadFile / form data
python-dotenv>=1.0.0             # Load .env variables

# ==================== VALIDATION ====================
pydantic[email]>=2.0.0           # EmailStr for user schemas
pydantic-settings>=2.0.0         # BaseSettings + .env loading

# ==================== DATABASE ====================
sqlalchemy>=2.0.0,<3.0.0         # Async ORM (DeclarativeBase style)
asyncpg>=0.29.0                  # PostgreSQL async driver (required for Railway/Neon)
aiosqlite>=0.19.0                # Async SQLite driver

# ==================== AUTH & SECURITY ====================
python-jose[cryptography]>=3.3.0 # JWT token encode/decode
passlib[argon2]>=1.7.4           # Password hashing context
argon2-cffi>=21.0.0              # Argon2 backend for passlib

# ==================== AI ====================
google-genai>=1.0.0              # Gemini 2.5 Pro API (text + vision)

# ==================== PDF PROCESSING ====================
PyMuPDF>=1.23.0                  # Primary: text extraction + PDF→image (vision mode)
pdfplumber>=0.10.0               # Fallback: good for tables
pypdf>=3.0.0                     # Last resort fallback
# pdf2image>=1.16.0              # Optional fallback (needs poppler system lib)

# ==================== DOCUMENT PROCESSING ====================
python-docx>=1.0.0               # DOCX text extraction + export
lxml>=4.9.0                      # XML/OMML processing (LaTeX→DOCX math)

# ==================== OCR (OPTIONAL) ====================
# pix2text[multilingual]>=1.1.4  # Local OCR + math formula recognition (LaTeX)
# magic-pdf[full]                # MinerU: layout-aware PDF extraction (heavy, optional)

# ==================== VECTOR SEARCH ====================
numpy>=1.24.0                    # Vectorized cosine similarity (SQLite fallback)
pgvector>=0.3.0                  # PostgreSQL vector similarity (pgvector extension)